Checks for existing files

+++ src/download.py
@@ -2,6 +2,7 @@
 
 import requests
 import re
+import os
 
 periode = 19
 base_url = 'https://www.bundestag.de'
@@ -32,6 +33,7 @@ def getListXML():
 
 def downloadXMLs(dw_list):
     for xmlfile in dw_list:
+        if os.path.isfile('data/pp19-data/' + xmlfile[-14:]): break
         print('Downloading:', xmlfile)
         with open('data/pp19-data/' + xmlfile[-14:], 'wb') as fp:
             fp.write(requests.get(xmlfile).content)
